Split bindings into focused files
Split the monolithic bindings.go into API-family files:

1) Put shared declarations in enums.go, types.go, and handles.go.
2) Put setup and lifecycle APIs in connection.go, config.go, and errors.go.
3) Put Query, result helpers, and chunk fetching in query.go.
4) Put prepared, value, logical type, and data chunk APIs in matching files.
5) Keep vector.go and selection_vector.go as separate lookup targets.
6) Put appender, table description, scalar function, and table function APIs
   in matching files.
7) Put Arrow, logging, profiling, replacement scan, and expression APIs in
   matching files.
8) Merge grab-bags into owners: allocation helpers and counters into alloc.go,
   scalar conversions and StringT helpers into types.go, UTF-8 validation into
   value.go, and stream result helpers into query.go.
9) Rename tests and Arrow files to drop the bindings_ prefix.
10) Remove section marker comments and TODO-only roadmaps for fully
    unimplemented API families while preserving API doc comments.
