+++ core_test.go
